+++ web/ui/scale-controller.js
@@ -247,6 +247,8 @@ const ScaleController = (() => {
     toggleAuto,
     toggleLog,
     syncUI,
+    isPointerOnPriceScale,
+    priceScaleWidth,
   };
 })();